+++ models/core/src/events/shard.rs
@@ -5,6 +5,8 @@ use super::Event;
 use super::EventBuilder;
 use super::Payload;
 use crate::agent::Shard;
+use crate::scope::EntityId;
+use crate::scope::Namespace;
 
 /// Metadata attached to shard allocation changed events.
 #[derive(Clone, Eq, PartialEq, Hash, Debug, Serialize, Deserialize)]
@@ -31,15 +33,6 @@ pub enum ShardEvent {
 }
 
 impl ShardEvent {
-    /// Look up the cluster ID for the event, if they have one.
-    pub fn cluster_id(&self) -> Option<&str> {
-        let cluster_id = match self {
-            ShardEvent::AllocationChanged(change) => &change.cluster_id,
-            ShardEvent::AllocationNew(shard) => &shard.cluster_id,
-        };
-        Some(cluster_id)
-    }
-
     /// Returns the event "code", the string that represents the event type.
     pub fn code(&self) -> &'static str {
         match self {
@@ -48,9 +41,23 @@ impl ShardEvent {
         }
     }
 
-    /// Returns the "ordering ID" for correctly streaming the event.
-    pub fn stream_key(&self) -> &str {
-        self.cluster_id().unwrap_or("<system>")
+    /// Identifier of the cluster the node event is about.
+    pub fn entity_id(&self) -> EntityId {
+        let cluster_id = match self {
+            ShardEvent::AllocationChanged(change) => &change.cluster_id,
+            ShardEvent::AllocationNew(shard) => &shard.cluster_id,
+        };
+        let node = match self {
+            ShardEvent::AllocationChanged(change) => &change.node_id,
+            ShardEvent::AllocationNew(shard) => &shard.node_id,
+        };
+        let shard = match self {
+            ShardEvent::AllocationChanged(change) => &change.shard_id,
+            ShardEvent::AllocationNew(shard) => &shard.shard_id,
+        };
+        // TODO: Must use a static string because of refs until nodes have namespaces attached.
+        let _ns = Namespace::HARDCODED_FOR_ROLLOUT();
+        EntityId::Shard("default", cluster_id, node, shard)
     }
 }